backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
librosa==0.10.2.post1
soundfile==0.12.1
numpy==1.26.4
transformers==4.40.0
torch==2.2.2
accelerate==0.29.3
python-dotenv==1.0.1
lingodotdev==1.3.0
